refactor(database): 将数据库相关代码提取到单独模块
重构数据库相关代码，将其从主模块分离到单独的database.py文件中
提高代码可维护性和模块化程度
